+++ apps/api/src/common/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './anonymize-ip'
+export * from './redact-email'
 export * from './sanitize-headers'